Overwrite function

+++ python/scan.py
@@ -26,7 +26,8 @@ class Scan:
                  masses: 'Masses',
                  modelname,
                  decay,
-                 maxwidth):
+                 maxwidth,
+                 overwrite=False):
 
         # store model name
         self.modelname = modelname
@@ -55,15 +56,15 @@ class Scan:
         # directory where we want the output to go
         self.outdir = fileutils.scanDir(modelname=modelname,decay=decay,masses=masses)
 
-        # remove previous directory if set to overwrite
-        if os.path.exists(self.outdir):
+         # remove previous directory if set to overwrite
+        if os.path.exists(self.outdir) and overwrite:
+            # remove directory
             shutil.rmtree(self.outdir)
 
-        # make working directory
-        os.makedirs(self.outdir)
-
-        # directory to store all of the output files
-        os.makedirs(self.outdir+"/files")
+        # check if directory exists, if not make it
+        if not os.path.exists(self.outdir):
+            os.makedirs(self.outdir)
+            os.makedirs(self.outdir+"/files")
 
         # create summary file
         self.summaryname = self.outdir+"scansummary_"+self.modelname+"_"+self.decay+"_"+str(self.masses)+".txt"
@@ -480,6 +481,7 @@ if __name__ == "__main__":
     argparser.add_argument("-r", "--parameter_rate", default=0.05, type=float, help="Rate at which parameter range should shrink")
     argparser.add_argument("-g", "--density_growth", default=0.2, type=float, help="Rate at which point density should grow")
     argparser.add_argument("-m", "--multiprocessing", action="store_true", help="Whether multiprocessing should be used")
+    argparser.add_argument("-o", "--overwrite", action="store_true", help="Whether overwrite should be used")
     args = argparser.parse_args()
 
     # create masses object
@@ -493,7 +495,8 @@ if __name__ == "__main__":
     myScan = Scan(masses=masses,
                   modelname=args.model,
                   decay=args.decay,
-                  maxwidth=args.maxwidth)
+                  maxwidth=args.maxwidth,
+                  overwrite=args.overwrite)
     
     # run scan using scan object
     myScan.runScan(npoints=args.npoints,
